+++ src/components/admin/ListChannels.vue
@@ -1,44 +1,43 @@
 <template lang="pug">
   .ListChannels
 
-    .columns
-      .column
-        h1.title.
-          Channels
-        h2.subtitle.
-          Channels description
-
-    .columns
-      .column
-        .box
-          table.table.is-striped.is-hoverable.is-fullwidth()
-            thead
-              tr
-                th.
-                  Channels Header 1
-                th.
-                  Channels Header 2
-                th.
-                  Actions
-            tbody
-              tr
-                td.
-                  content 1
-                td.
-                  content 2
-                td
-                  button.button.is-info.is-small.is-outlined()
-                    span.
-                      Edit
-                  .button.is-danger.is-small.is-outlined()
-                    span.
-                      Delete
+    admin-table(
+      :itemPlural='itemPlural',
+      :itemSingular='itemSingular',
+      :routeAddName='routeAddName',
+      :routeEditName='routeEditName',
+      :storeGetter='storeGetter',
+      :storeRemove='storeRemove',
+      :subtitle='subtitle',
+      :tableCols='tableCols',
+      :title='title',
+      )
 </template>
 
 <script>
+import AdminTable from '@/components/admin/ui/AdminTable'
 
 export default {
   name: 'ListChannels',
+  components: {
+    AdminTable,
+  },
+  data () {
+    return {
+      itemPlural: 'Channels',
+      itemSingular: 'Channel',
+      routeAddName: 'ChannelAdd',
+      routeEditName: 'ChannelEdit',
+      storeGetter: 'channelsList',
+      storeRemove: 'channelRemove',
+      subtitle: 'Channels description',
+      tableCols: [
+        { label: 'Name', prop: 'name' },
+        { label: 'Type', prop: 'type' },
+      ],
+      title: 'Channels',
+    }
+  },
 }
 </script>